[WOR-431] Adjust azure profile data model managed app coordinates (#15)
Context

There are a couple of problems with the current create profile code path with respect to Azure.
It pulls the app down for access control checks by using the application deployment name and the parent resource group name. This is problematic because we:

Should be using the same code path for access control checks when enumerating managed apps via the /managedApps endpoint as we are when creating a profile
We need the managed resource group ID to be persisted with the profile so we can provide that to WSM downstream when creating workspace resources.

This PR

Refactors the data model to swap out applicationDeploymentName + resourceGroupName for managedResourceGroupId
Changes the profile creation logic to reuse the managed application enumeration codepath in AzureService as the /managedApps endpoint uses
Introduces a BaseUnitTest base class for use when testing POJOs. In my testing, it was becoming quite cumbersome/slow to spin up spring boot tests for basic business logic tests.