[WOR-454] Reworking well-known offers (#25)
Context

The way "well-known" Azure marketplace offers is configured in BPM is awkward and difficult to modify via helmfiles.

(Quick refresher: well known offers are those Azure application offers whose deployments BPM will allow to be linked to billing profiles)

Specifically, having the offer name be a key and the auth'ed user key be a value makes setting the config via OS environment variables intractable in my research.

This PR

Restructures how we configure well known application offers. The config is now an array of offer objects, with the keys being the offer name and the authed user key. The knock-on effect here is that we are able to set the offers via environment variables like so:
PROFILE_AZURE_APPLICATIONOFFERS_0_NAME=foo;PROFILE_AZURE_APPLICATIONOFFERS_0_AUTHORIZEDUSERKEY=bar
or via a helmfile change