[WOR-424] Validate Azure resource providers (#45)
Context

We need to check that needed Azure resource providers are registered in a subscription before creating a billing profile, otherwise Terra will be non-functional.

This PR

Adds a check in the verification of the MRG step during profile creation that ensures all required providers are registered. If any are missing, we fail out with a 409.
We talked about using 424, but it looks like that is a WebDAV specific error code according to MDN.
Adds the required providers to the application.yaml configuration.