[WOR-515] Return 403 instead of 401 when user lacks access to a profile (#61)
We are returning 401 UNAUTHORIZED when a user does not have access to a profile. We should return 403 FORBIDDEN as that is a more accurate status code -- 401 generally indicates a bad authentication credential, which is not the case here. 403 should be used to indicate the credential is good but the server is not allowing access to the desired resource.