[WOR-1535] Spring Boot dependency: pin postgresql to 42.7.2
Our current version of Spring Boot is 3.1.2, which brings in postgresql 42.6.0.  To mitigate a security vulnerability, we must instruct Spring Boot use a non-vulnerable version.
I first tried upgrading our TCL but this had no impact: the Spring Boot pin was both necessary and sufficient, so I will pursue the TCL upgrade separately as it's a larger change.